rm tokenize_spans and reuse scratch pool for pre-tokens
